pythonGRPC/pyproject.toml
[project]
name = "generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.12",
    "feedparser>=6.0.11",
    "grpcio>=1.73.0",
    "grpcio-tools>=1.73.0",
    "lxml-html-clean>=0.4.2",
    "newspaper3k>=0.2.8",
    "nltk>=3.9.1",
    "rss-parser>=2.1.0",
    "sentence-transformers>=4.1.0",
    "text-prettifier>=2.0.1",
    "yfinance>=0.2.63",
]
